Separated each classes used by spectral_indicies into two files Cmd.py and ScriptHelper.py (also contains ScriptArgParser). Updated Makefiles to have correct comments
